+++ common/src/main/aidl/com/microsoft/identity/client/IMicrosoftAuthService.aidl
@@ -21,7 +21,21 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package com.microsoft.identity.common.internal.broker;
+package com.microsoft.identity.client;
+/**
+ * Broker Account service APIs provided by the broker app. Those APIs will be responsible for interacting with the
+ * account manager API. Calling app does not need to request for contacts permission if the broker installed on the
+ * device has the support for the bound service.
+ */
+interface IMicrosoftAuthService {
 
-// Parcelable for BrokerRequest
-parcelable BrokerRequest;
+    Bundle connect(in Bundle bundle);
+
+    Bundle getAccounts();
+
+    Bundle acquireTokenSilently(in Bundle requestBundle);
+
+    Intent getIntentForInteractiveRequest();
+
+    void removeAccounts();
+}
