+++ common/src/main/java/com/microsoft/identity/common/internal/broker/BrokerRequest.java
@@ -27,10 +27,12 @@ import android.os.Parcelable;
 
 import com.google.gson.annotations.SerializedName;
 
+import java.io.Serializable;
+
 /**
  * Represents the broker request
  */
-public class BrokerRequest implements Parcelable {
+public class BrokerRequest implements Serializable {
 
     private class SerializedNames {
         public final static String AUTHORITY = "account.authority";
@@ -46,12 +48,10 @@ public class BrokerRequest implements Parcelable {
         public final static String AUTHORITY_TYPE = "account.authorityType";
         public final static String EXTRA_QUERY_STRING_PARAMETER = "account.extra.query.param";
         public final static String CORRELATION_ID = "account.correlationid";
-        public final static String NAME = "account.name";
         public final static String PROMPT = "account.prompt";
         public final static String CLAIMS = "account.claims";
         public final static String FORCE_REFRESH = "force.refresh";
         public final static String APPLICATION_NAME = "application.name";
-        public final static String APPLICATION_UID = "application.uid";
         public final static String APPLICATION_VERSION = "application.version";
         public final static String MSAL_VERSION_KEY = "msal.version";
         public final static String EXPIRATION_BUFFER = "expiration.buffer";
@@ -96,9 +96,6 @@ public class BrokerRequest implements Parcelable {
     @SerializedName(SerializedNames.CORRELATION_ID)
     private String mCorrelationId;
 
-    @SerializedName(SerializedNames.NAME)
-    private String mName;
-
     @SerializedName(SerializedNames.PROMPT)
     private String mPrompt;
 
@@ -110,9 +107,6 @@ public class BrokerRequest implements Parcelable {
 
     @SerializedName(SerializedNames.APPLICATION_NAME)
     private String mApplicationName;
-
-    @SerializedName(SerializedNames.APPLICATION_UID)
-    private int mUId;
 
     @SerializedName((SerializedNames.APPLICATION_VERSION))
     private String mApplicationVersion;
@@ -127,76 +121,6 @@ public class BrokerRequest implements Parcelable {
     public BrokerRequest() {
 
     }
-
-    protected BrokerRequest(Parcel in) {
-        mAuthority = in.readString();
-        mScope = in.readString();
-        mRedirect = in.readString();
-        mClientId = in.readString();
-        mEnvironment = in.readString();
-        mRealm = in.readString();
-        mHomeAccountId = in.readString();
-        mLocalAccountId = in.readString();
-        mUserName = in.readString();
-        mClientInfo = in.readString();
-        mAuthorityType = in.readString();
-        mExtraQueryStringParameter = in.readString();
-        mCorrelationId = in.readString();
-        mName = in.readString();
-        mPrompt = in.readString();
-        mClaims = in.readString();
-        int tmpMForceRefresh = in.readInt();
-        mForceRefresh = tmpMForceRefresh != 0;
-        mApplicationName = in.readString();
-        mUId = in.readInt();
-        mApplicationVersion = in.readString();
-        mMsalVersion = in.readString();
-        mExpirationBuffer = in.readInt();
-    }
-
-    public static final Creator<BrokerRequest> CREATOR = new Creator<BrokerRequest>() {
-        @Override
-        public BrokerRequest createFromParcel(Parcel in) {
-            return new BrokerRequest(in);
-        }
-
-        @Override
-        public BrokerRequest[] newArray(int size) {
-            return new BrokerRequest[size];
-        }
-    };
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeString(mAuthority);
-        dest.writeString(mScope);
-        dest.writeString(mRedirect);
-        dest.writeString(mClientId);
-        dest.writeString(mEnvironment);
-        dest.writeString(mRealm);
-        dest.writeString(mHomeAccountId);
-        dest.writeString(mLocalAccountId);
-        dest.writeString(mUserName);
-        dest.writeString(mClientInfo);
-        dest.writeString(mAuthorityType);
-        dest.writeString(mExtraQueryStringParameter);
-        dest.writeString(mCorrelationId);
-        dest.writeString(mName);
-        dest.writeString(mPrompt);
-        dest.writeString(mClaims);
-        dest.writeInt(mForceRefresh ? 1 : 0);
-        dest.writeString(mApplicationName);
-        dest.writeInt(mUId);
-        dest.writeString(mApplicationVersion);
-        dest.writeString(mMsalVersion);
-        dest.writeInt(mExpirationBuffer);
-    }
-
 
     public String getAuthority() {
         return mAuthority;
@@ -303,14 +227,6 @@ public class BrokerRequest implements Parcelable {
         this.mCorrelationId = correlationId;
     }
 
-    public String getName() {
-        return mName;
-    }
-
-    public void setName(String name) {
-        this.mName = name;
-    }
-
     public String getPrompt() {
         return mPrompt;
     }
@@ -341,14 +257,6 @@ public class BrokerRequest implements Parcelable {
 
     public void setApplicationName(String applicationName) {
         this.mApplicationName = applicationName;
-    }
-
-    public int getUId() {
-        return mUId;
-    }
-
-    public void setUId(int uId) {
-        this.mUId = uId;
     }
 
     public String getApplicationVersion() {
